multi-video-upload/requirements.txt
google-auth==2.23.0
google-auth-oauthlib==0.4.1
google-auth-httplib2==0.0.3
google-api-python-client==1.12.5
instagram-private-api==2.3.0
pytiktok
moviepy==1.0.3